Trait llvm_ir::debugloc::HasDebugLoc

pub trait HasDebugLoc {
    fn get_debug_loc(&self) -> &Option<DebugLoc>;
}

Required methods

fn get_debug_loc(&self) -> &Option<DebugLoc>

Returns the DebugLoc associated with the given Instruction, Terminator, GlobalVariable, or Function; or None if it doesn't have a DebugLoc.

Reasons something might not have a DebugLoc include: (1) the file was compiled without debuginfo; (2) for an Instruction, it might not directly correspond to any source line. For instance, it may be just setting up the stack frame for a function.

Loading content...