Module llvm_ir::instruction

Modules

groups

Structs

AShr
Arithmetic shift right. See LLVM 8 docs on the 'ashr' instruction

Add
See LLVM 8 docs on the 'add' instruction

AddrSpaceCast
See LLVM 8 docs on the 'addrspacecast' instruction

Alloca
Allocate memory on the stack. See LLVM 8 docs on the 'alloca' instruction

And
Bitwise logical and. See LLVM 8 docs on the 'and' instruction

AtomicRMW
Atomic read-modify-write. See LLVM 8 docs on the 'atomicrmw' instruction

Atomicity
See LLVM 8 docs on Atomic Memory Ordering Constraints

BitCast
Convert between types without changing any bits. See LLVM 8 docs on the 'bitcast' instruction

Call
Function call. See LLVM 8 docs on the 'call' instruction

CatchPad
Used for exception handling. See LLVM 8 docs on the 'catchpad' instruction

CleanupPad
Used for exception handling. See LLVM 8 docs on the 'cleanuppad' instruction

CmpXchg
Atomic compare and exchange. See LLVM 8 docs on the 'cmpxchg' instruction

ExtractElement
Get an element from a vector at a specified index. See LLVM 8 docs on the 'extractelement' instruction

ExtractValue
Extract the value of a member field from an aggregate (struct or array) type. See LLVM 8 docs on the 'extractvalue' instruction

FAdd
Floating-point add. See LLVM 8 docs on the 'fadd' instruction

FCmp
Compare floating-point values or vectors of floating-point values. See LLVM 8 docs on the 'fcmp' instruction

FDiv
Floating-point divide. See LLVM 8 docs on the 'fdiv' instruction

FMul
Floating-point multiply. See LLVM 8 docs on the 'fmul' instruction

FNeg
Floating-point unary negation. See LLVM 8 docs on the 'fneg' instruction

FPExt
Extend a floating-point value. See LLVM 8 docs on the 'fpext' instruction

FPToSI
Convert floating-point to signed integer. See LLVM 8 docs on the 'fptosi' instruction

FPToUI
Convert floating-point to unsigned integer. See LLVM 8 docs on the 'fptoui' instruction

FPTrunc
Truncate a floating-point value. See LLVM 8 docs on the 'fptrunc' instruction

FRem
Floating-point remainder. See LLVM 8 docs on the 'frem' instruction

FSub
Floating-point sub. See LLVM 8 docs on the 'fsub' instruction

FastMathFlags
See LLVM 8 docs on Fast-Math Flags

Fence
See LLVM 8 docs on the 'fence' instruction

GetElementPtr
Get the address of a subelement of an aggregate data structure. Only performs address calculation, does not actually access memory. See LLVM 8 docs on the 'getelementptr' instruction

ICmp
Compare integers, pointers, or vectors of integers or pointers. See LLVM 8 docs on the 'icmp' instruction

InlineAssembly
See LLVM 8 docs on Inline Assembler Expressions. InlineAssembly needs more fields, but the necessary getter functions are apparently not exposed in the LLVM C API (only the C++ API)

InsertElement
Insert an element into a vector at a specified index. See LLVM 8 docs on the 'insertelement' instruction

InsertValue
Insert a value into a member field of an aggregate (struct or array) type. See LLVM 8 docs on the 'insertvalue' instruction

IntToPtr
Convert integer to pointer. See LLVM 8 docs on the 'inttoptr' instruction

LShr
Logical shift right. See LLVM 8 docs on the 'lshr' instruction

LandingPad
Used for exception handling. See LLVM 8 docs on the 'landingpad' instruction

LandingPadClause
LandingPadClause needs more fields, but the necessary getter functions are apparently not exposed in the LLVM C API (only the C++ API)

Load
Load a value from memory. See LLVM 8 docs on the 'load' instruction

Mul
See LLVM 8 docs on the 'mul' instruction

Or
Bitwise logical inclusive or. See LLVM 8 docs on the 'or' instruction

Phi
See LLVM 8 docs on the 'phi' instruction

PtrToInt
Convert pointer to integer. See LLVM 8 docs on the 'ptrtoint' instruction

SDiv
See LLVM 8 docs on the 'sdiv' instruction

SExt
Sign-extend. See LLVM 8 docs on the 'sext' instruction

SIToFP
Convert signed integer to floating-point. See LLVM 8 docs on the 'sitofp' instruction

SRem
See LLVM 8 docs on the 'srem' instruction

Select
See LLVM 8 docs on the 'select' instruction

Shl
Shift left. See LLVM 8 docs on the 'shl' instruction

ShuffleVector
See LLVM 8 docs on the 'shufflevector' instruction

Store
Store a value to memory. See LLVM 8 docs on the 'store' instruction

Sub
See LLVM 8 docs on the 'sub' instruction

Trunc
Truncate. See LLVM 8 docs on the 'trunc' instruction

UDiv
See LLVM 8 docs on the 'udiv' instruction

UIToFP
Convert unsigned integer to floating-point. See LLVM 8 docs on the 'uitofp' instruction

URem
See LLVM 8 docs on the 'urem' instruction

VAArg
See LLVM 8 docs on the 'va_arg' instruction

Xor
Bitwise logical exclusive or. See LLVM 8 docs on the 'xor' instruction

ZExt
Zero-extend. See LLVM 8 docs on the 'zext' instruction

Enums

Instruction
Non-terminator instructions.

MemoryOrdering
See LLVM 8 docs on Atomic Memory Ordering Constraints

RMWOperation
See LLVM 8 docs on the 'atomicrmw' instruction

SynchronizationScope
See LLVM 8 docs on Atomic Memory Ordering Constraints

TailCallKind

Traits

BinaryOp

HasResult

UnaryOp