Enum llvm_ir::instruction::Instruction

pub enum Instruction {
    Add(Add),
    Sub(Sub),
    Mul(Mul),
    UDiv(UDiv),
    SDiv(SDiv),
    URem(URem),
    SRem(SRem),
    And(And),
    Or(Or),
    Xor(Xor),
    Shl(Shl),
    LShr(LShr),
    AShr(AShr),
    FAdd(FAdd),
    FSub(FSub),
    FMul(FMul),
    FDiv(FDiv),
    FRem(FRem),
    FNeg(FNeg),
    ExtractElement(ExtractElement),
    InsertElement(InsertElement),
    ShuffleVector(ShuffleVector),
    ExtractValue(ExtractValue),
    InsertValue(InsertValue),
    Alloca(Alloca),
    Load(Load),
    Store(Store),
    Fence(Fence),
    CmpXchg(CmpXchg),
    AtomicRMW(AtomicRMW),
    GetElementPtr(GetElementPtr),
    Trunc(Trunc),
    ZExt(ZExt),
    SExt(SExt),
    FPTrunc(FPTrunc),
    FPExt(FPExt),
    FPToUI(FPToUI),
    FPToSI(FPToSI),
    UIToFP(UIToFP),
    SIToFP(SIToFP),
    PtrToInt(PtrToInt),
    IntToPtr(IntToPtr),
    BitCast(BitCast),
    AddrSpaceCast(AddrSpaceCast),
    ICmp(ICmp),
    FCmp(FCmp),
    Phi(Phi),
    Select(Select),
    Freeze(Freeze),
    Call(Call),
    VAArg(VAArg),
    LandingPad(LandingPad),
    CatchPad(CatchPad),
    CleanupPad(CleanupPad),
}

Non-terminator instructions.

Variants

Add(Add)

Sub(Sub)

Mul(Mul)

UDiv(UDiv)

SDiv(SDiv)

URem(URem)

SRem(SRem)

And(And)

Or(Or)

Xor(Xor)

Shl(Shl)

LShr(LShr)

AShr(AShr)

FAdd(FAdd)

FSub(FSub)

FMul(FMul)

FDiv(FDiv)

FRem(FRem)

FNeg(FNeg)

ExtractElement(ExtractElement)

InsertElement(InsertElement)

ShuffleVector(ShuffleVector)

ExtractValue(ExtractValue)

InsertValue(InsertValue)

Alloca(Alloca)

Load(Load)

Store(Store)

Fence(Fence)

CmpXchg(CmpXchg)

AtomicRMW(AtomicRMW)

GetElementPtr(GetElementPtr)

Trunc(Trunc)

ZExt(ZExt)

SExt(SExt)

FPTrunc(FPTrunc)

FPExt(FPExt)

FPToUI(FPToUI)

FPToSI(FPToSI)

UIToFP(UIToFP)

SIToFP(SIToFP)

PtrToInt(PtrToInt)

IntToPtr(IntToPtr)

BitCast(BitCast)

AddrSpaceCast(AddrSpaceCast)

ICmp(ICmp)

FCmp(FCmp)

Phi(Phi)

Select(Select)

Freeze(Freeze)

Call(Call)

VAArg(VAArg)

LandingPad(LandingPad)

CatchPad(CatchPad)

CleanupPad(CleanupPad)

Implementations

impl Instruction

pub fn try_get_result(&self) -> Option<&Name>

Get the result (destination) of the Instruction, or None if the Instruction doesn't have a result (has void type).

pub fn is_atomic(&self) -> bool

Whether the Instruction is atomic

impl Instruction

pub fn is_binary_op(&self) -> bool

Determine if the Instruction is one of the ones in groups::BinaryOp, without actually using try_into() (which would consume it)

pub fn is_unary_op(&self) -> bool

Determine if the Instruction is one of the ones in groups::UnaryOp, without actually using try_into() (which would consume it)