Module llvm_ir::terminator

Structs

Br
See LLVM 10 docs on the 'br' instruction. The LLVM 'br' instruction has both conditional and unconditional variants, which we separate -- this is the unconditional variant, while the conditional variant is CondBr.

CallBr
See LLVM 10 docs on the 'callbr' instruction

CatchRet
See LLVM 10 docs on the 'catchret' instruction

CatchSwitch
See LLVM 10 docs on the 'catchswitch' instruction

CleanupRet
See LLVM 10 docs on the 'cleanupret' instruction

CondBr
See LLVM 10 docs on the 'br' instruction. The LLVM 'br' instruction has both conditional and unconditional variants, which we separate -- this is the conditional variant, while the unconditional variant is Br.

IndirectBr
See LLVM 10 docs on the 'indirectbr' instruction

Invoke
See LLVM 10 docs on the 'invoke' instruction

Resume
See LLVM 10 docs on the 'resume' instruction

Ret
See LLVM 10 docs on the 'ret' instruction

Switch
See LLVM 10 docs on the 'switch' instruction

Unreachable
See LLVM 10 docs on the 'unreachable' instruction

Enums

Terminator
Terminator instructions end a basic block. See LLVM 10 docs on Terminator Instructions