Enum llvm_ir::types::Type

pub enum Type {
    VoidType,
    IntegerType {
        bits: u32,
    },
    PointerType {
        pointee_type: Box<Type>,
        addr_space: AddrSpace,
    },
    FPType(FPType),
    FuncType {
        result_type: Box<Type>,
        param_types: Vec<Type>,
        is_var_arg: bool,
    },
    VectorType {
        element_type: Box<Type>,
        num_elements: usize,
    },
    ArrayType {
        element_type: Box<Type>,
        num_elements: usize,
    },
    StructType {
        element_types: Vec<Type>,
        is_packed: bool,
    },
    NamedStructType {
        name: String,
        ty: Option<Weak<RwLock<Type>>>,
    },
    X86_MMXType,
    MetadataType,
    LabelType,
    TokenType,
}

See LLVM 9 docs on Type System

Variants

VoidType

See LLVM 9 docs on Void Type

IntegerType

See LLVM 9 docs on Integer Type

Fields of IntegerType

bits: u32

PointerType

See LLVM 9 docs on Pointer Type

Fields of PointerType

pointee_type: Box<Type>addr_space: AddrSpace

FPType(FPType)

See LLVM 9 docs on Floating-Point Types

FuncType

See LLVM 9 docs on Function Type

Fields of FuncType

result_type: Box<Type>param_types: Vec<Type>is_var_arg: bool

VectorType

Vector types (along with integer, FP, pointer, and X86_MMX types) are "first class types", which means they can be produced by instructions (see LLVM 9 docs on First Class Types). See LLVM 9 docs on Vector Type

Fields of VectorType

element_type: Box<Type>num_elements: usize

ArrayType

Struct and Array types (but not vector types) are "aggregate types" and cannot be produced by a single instruction (see LLVM 9 docs on Aggregate Types). See LLVM 9 docs on Array Type

Fields of ArrayType

element_type: Box<Type>num_elements: usize

StructType

The StructType variant is used for a "literal" (i.e., anonymous) structure type. See LLVM 9 docs on Structure Type

Fields of StructType

element_types: Vec<Type>is_packed: bool

NamedStructType

Named structure types. Note that these may be self-referential (i.e., recursive). See LLVM 9 docs on Structure Type

Fields of NamedStructType

name: String

Name of the struct type

ty: Option<Weak<RwLock<Type>>>

The actual struct type, which will be a StructType variant. A None here indicates an opaque type; see LLVM 9 docs on Opaque Structure Types. The weak reference should remain valid for at least the lifetime of the Module in which the named struct type is defined.

X86_MMXType

See LLVM 9 docs on X86_MMX Type

MetadataType

See LLVM 9 docs on Metadata Type

LabelType

LabelType is the type of BasicBlock labels. See LLVM 9 docs on Label Type

TokenType

See LLVM 9 docs on Token Type

Methods

impl Type

pub fn bool() -> Type

pub fn i8() -> Type

pub fn i16() -> Type

pub fn i32() -> Type

pub fn i64() -> Type

pub fn half() -> Type

pub fn single() -> Type

pub fn double() -> Type

pub fn pointer_to(ty: Type) -> Type