Enum llvm_ir::constant::Constant

pub enum Constant {
    Int {
        bits: u32,
        value: u64,
    },
    Float(Float),
    Null(Type),
    AggregateZero(Type),
    Struct {
        name: Option<String>,
        values: Vec<Constant>,
        is_packed: bool,
    },
    Array {
        element_type: Type,
        elements: Vec<Constant>,
    },
    Vector(Vec<Constant>),
    Undef(Type),
    BlockAddress,
    GlobalReference {
        name: Name,
        ty: Type,
    },
    TokenNone,
    Add(Box<Add>),
    Sub(Box<Sub>),
    Mul(Box<Mul>),
    UDiv(Box<UDiv>),
    SDiv(Box<SDiv>),
    URem(Box<URem>),
    SRem(Box<SRem>),
    And(Box<And>),
    Or(Box<Or>),
    Xor(Box<Xor>),
    Shl(Box<Shl>),
    LShr(Box<LShr>),
    AShr(Box<AShr>),
    FAdd(Box<FAdd>),
    FSub(Box<FSub>),
    FMul(Box<FMul>),
    FDiv(Box<FDiv>),
    FRem(Box<FRem>),
    ExtractElement(Box<ExtractElement>),
    InsertElement(Box<InsertElement>),
    ShuffleVector(Box<ShuffleVector>),
    ExtractValue(Box<ExtractValue>),
    InsertValue(Box<InsertValue>),
    GetElementPtr(Box<GetElementPtr>),
    Trunc(Box<Trunc>),
    ZExt(Box<ZExt>),
    SExt(Box<SExt>),
    FPTrunc(Box<FPTrunc>),
    FPExt(Box<FPExt>),
    FPToUI(Box<FPToUI>),
    FPToSI(Box<FPToSI>),
    UIToFP(Box<UIToFP>),
    SIToFP(Box<SIToFP>),
    PtrToInt(Box<PtrToInt>),
    IntToPtr(Box<IntToPtr>),
    BitCast(Box<BitCast>),
    AddrSpaceCast(Box<AddrSpaceCast>),
    ICmp(Box<ICmp>),
    FCmp(Box<FCmp>),
    Select(Box<Select>),
}

See LLVM 9 docs on Constants. Constants can be either values, or expressions involving other constants (see LLVM 9 docs on Constant Expressions).

Variants

Int

Fields of Int

bits: u32value: u64

Float(Float)

Null(Type)

The Type here must be a PointerType. See LLVM 9 docs on Simple Constants

AggregateZero(Type)

A zero-initialized array or struct (or scalar).

Struct

Fields of Struct

name: Option<String>values: Vec<Constant>is_packed: bool

Array

Fields of Array

element_type: Typeelements: Vec<Constant>

Vector(Vec<Constant>)

Undef(Type)

Undef can be used anywhere a constant is expected. See LLVM 9 docs on Undefined Values

BlockAddress

The address of the given (non-entry) BasicBlock. See LLVM 9 docs on Addresses of Basic Blocks. BlockAddress needs more fields, but the necessary getter functions are apparently not exposed in the LLVM C API (only the C++ API)

GlobalReference

Fields of GlobalReference

name: Namety: Type

TokenNone

Add(Box<Add>)

Sub(Box<Sub>)

Mul(Box<Mul>)

UDiv(Box<UDiv>)

SDiv(Box<SDiv>)

URem(Box<URem>)

SRem(Box<SRem>)

And(Box<And>)

Or(Box<Or>)

Xor(Box<Xor>)

Shl(Box<Shl>)

LShr(Box<LShr>)

AShr(Box<AShr>)

FAdd(Box<FAdd>)

FSub(Box<FSub>)

FMul(Box<FMul>)

FDiv(Box<FDiv>)

FRem(Box<FRem>)

ExtractElement(Box<ExtractElement>)

InsertElement(Box<InsertElement>)

ShuffleVector(Box<ShuffleVector>)

ExtractValue(Box<ExtractValue>)

InsertValue(Box<InsertValue>)

GetElementPtr(Box<GetElementPtr>)

Trunc(Box<Trunc>)

ZExt(Box<ZExt>)

SExt(Box<SExt>)

FPTrunc(Box<FPTrunc>)

FPExt(Box<FPExt>)

FPToUI(Box<FPToUI>)

FPToSI(Box<FPToSI>)

UIToFP(Box<UIToFP>)

SIToFP(Box<SIToFP>)

PtrToInt(Box<PtrToInt>)

IntToPtr(Box<IntToPtr>)

BitCast(Box<BitCast>)

AddrSpaceCast(Box<AddrSpaceCast>)

ICmp(Box<ICmp>)

FCmp(Box<FCmp>)

Select(Box<Select>)